Add refactore code
