+++ api/__init__.py
@@ -9,7 +9,7 @@ def create_app():
     jwt = JWTManager(app)
 
     from api.views.user import user_blueprint as user_blueprint
-    from api.views.parcels import parcel_blueprint as parcel_blueprint
+    from api.views.orders import parcel_blueprint as parcel_blueprint
 
     app.register_blueprint(user_blueprint)
     app.register_blueprint(parcel_blueprint)
